lab2/pyproject.toml
[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "reinvent-bedrock-workshop"
version = "0.2.0"
description = "Bedrock & prompt engineering workshop for re:Invent-23"
authors = [
    "Philipp Kaindl <philikai@amazon.de>",
    "Akarsha Sehwag <akshseh@amazon.de>",
    "Olivier Boder <bodero@amazon.de>",
    "Vikesh Pandey <pandvike@amazon.co.uk>",
    "Tingyi Li <tingyity@amazon.com>", 
]
license = "Amazon Software License"

[tool.poetry.dependencies]
python = "~3.9.0"
aws-cdk-lib = ">=2.99.0"
constructs = "^10.0.0"
boto3 = "~1.27.0"
pyyaml = "^6.0.0"
aws-cdk-aws-apigatewayv2-alpha = "^2.0"
aws-cdk-aws-apigatewayv2-integrations-alpha = "^2.0.0"
aws-cdk-aws-apigatewayv2-authorizers-alpha = "^2.0.0"
requests = "~2.31.0"
aws-lambda-powertools = {extras = ["validation"], version = "^2.25.1"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
black = "^23.7.0"
ruff = "^0.0.282"
isort = "^5.12.0"
ipykernel = "^6.25.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.0.0"
hypothesis = "^6.87.3"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
known_first_party = ["utils", "model", "retriever", "messaging"]

[tool.ruff]
select = ["E", "F", "C", "W", "R", "B", "Q"]
line-length = 120
target-version="py39"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "venv",
    ".venv",
]

[tool.black]
target-version = ['py39']
line-length = 120

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
    ".",    # Adds the project's root directory and makes the `infra` package importable
    "./assets/layers",
    "./assets/lambda",
]
norecursedirs = [
    "cdk.out"
]
